requirements.txt
aiogram==3.13.1
flask==3.0.3
gunicorn==23.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
werkzeug==3.0.6